requirements.txt
openai-whisper>=1.1.10
torch>=2.0.0
tqdm>=4.65.0
numpy>=1.24.0
colorama>=0.4.4
psutil>=5.9.0
flask>=2.0.0
flask-wtf>=1.0.0
